# swagger-spec.json
{"openapi":"3.0.0","paths":{"/user/update":{"put":{"operationId":"UserController_update","summary":"update one user","parameters":[{"schema":{}}],"requestBody":{"required":false,"content":{"multipart/form-data":{"schema":{"allOf":[{"$ref":"#/components/schemas/UpdateDto"},{"properties":{"avatar":{"type":"string","format":"binary"}},"type":"object"}]}}}},"responses":{"201":{"description":"success"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"},"404":{"description":"failed"}},"tags":["User"],"security":[{"bearer":[]}]}},"/file-upload/uploadToS3":{"post":{"operationId":"FileUploadController_create","summary":"The key name must be file","parameters":[],"responses":{"201":{"description":"success"},"401":{"description":"Unauthorized"},"500":{"description":"failed"}},"tags":["FileUpload"]}},"/file-upload/uploadManyToS3":{"post":{"operationId":"FileUploadController_uploadManyToS3","parameters":[],"responses":{"201":{"description":""}},"tags":["FileUpload"]}},"/file-upload/getFile":{"get":{"operationId":"FileUploadController_getFile","summary":"Get file from s3","parameters":[{"name":"key","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"success"},"400":{"description":"failed, the key not exist"},"401":{"description":"Unauthorized"}},"tags":["FileUpload"]}},"/file-upload/delete":{"delete":{"operationId":"FileUploadController_deleteFile","summary":"The key name must be file","parameters":[{"name":"key","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"success"},"400":{"description":"failed, the key not exist"},"401":{"description":"Unauthorized"}},"tags":["FileUpload"]}},"/items/save":{"post":{"operationId":"ItemsController_save","parameters":[{"schema":{}}],"requestBody":{"required":true,"description":"Items details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Items"}}}},"responses":{"201":{"description":"success"}},"tags":["Items"],"security":[{"bearer":[]}]}},"/items/update":{"put":{"operationId":"ItemsController_update","parameters":[{"schema":{}}],"requestBody":{"required":true,"description":"Items details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateItemDto"}}}},"responses":{"201":{"description":"success"}},"tags":["Items"],"security":[{"bearer":[]}]}}},"info":{"title":"Little Big Shed","description":"The API description","version":"1.0","contact":{}},"tags":[{"name":"User","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UpdateDto":{"type":"object","properties":{"fullName":{"type":"string","description":"fullName","example":"Andrew Owen"},"email":{"type":"string","example":"tester1@gmail.com"},"avatar":{"type":"string"},"mobile":{"type":"string","example":"0450123456","description":"Must be Australia mobile"},"address":{"type":"string","description":"address"},"lat":{"type":"number","description":"lat"},"lng":{"type":"number","description":"lng"},"suburb":{"type":"string","description":"suburb"},"bsb":{"type":"string","description":"bsb"},"account_number":{"type":"string","description":"account_number"},"lender_rating":{"type":"number","example":5,"description":"rating score as a lender."},"borrower_rating":{"type":"number","example":5,"description":"rating score as a borrower"},"available":{"type":"string","example":"11010111001100","description":"will set default to all 1"},"password":{"type":"string","description":"Password, length: 8-20, can be symbol, number, character mixed, cannot be pure number or character, like [\"abcd1234@\", \"abcd1234\"]","example":"abcd1234@"}},"required":["password"]},"Items":{"type":"object","properties":{"i_id":{"type":"number","description":"item ID"},"u_id":{"type":"string","description":"User ID"},"title":{"type":"string","description":"title"},"category":{"type":"string","description":"category"},"pictures":{"type":"string","description":"pictures, maximum is 9"},"description":{"type":"string","description":"Description"},"price":{"type":"number","description":"Price"},"deliveryPrice":{"type":"number","description":"delivery Price"},"rating":{"type":"number","description":"rating score"},"discount":{"type":"number","description":"discount for off peak"},"available":{"type":"string","example":"11010111001100","description":"will set default to all 1"},"lat":{"type":"number","description":"lat"},"lng":{"type":"number","description":"lng"},"address":{"type":"string","description":"address"},"suburb":{"type":"string","description":"suburb"},"created":{"format":"date-time","type":"string"},"updated":{"format":"date-time","type":"string"}},"required":["u_id","title","category","description","price","lat","lng"]},"UpdateItemDto":{"type":"object","properties":{"u_id":{"type":"string","description":"User ID"},"title":{"type":"string","description":"title"},"category":{"type":"string","description":"category"},"pictures":{"type":"string","description":"pictures, maximum is 9"},"description":{"type":"string","description":"Description"},"price":{"type":"number","description":"Price"},"deliveryPrice":{"type":"number","description":"delivery Price"},"rating":{"type":"number","description":"rating score"},"discount":{"type":"number","description":"discount for off peak"},"available":{"type":"string","example":"11010111001100","description":"will set default to all 1"},"lat":{"type":"number","description":"lat"},"lng":{"type":"number","description":"lng"},"address":{"type":"string","description":"address"},"suburb":{"type":"string","description":"suburb"},"deletedImages":{"type":"string","description":"images be deleted, separate by comma, image1,image2"},"newImages":{"description":"new file","type":"array","items":{"type":"string"}},"i_id":{"type":"number","description":"item ID, must provide"}},"required":["i_id"]}}}}